update prompts; etc

+++ src/imggen/vlm.py
@@ -1,5 +1,6 @@
 from pathlib import Path
 
+import torch
 from PIL import Image
 from qwen_vl_utils import process_vision_info
 from transformers import AutoModelForImageTextToText, AutoProcessor
@@ -78,7 +79,9 @@ def prepare_vlm_batch(
 
 def load_model(model_name: str, device: str):
     """Load VLM and processor."""
-    model = AutoModelForImageTextToText.from_pretrained(model_name, device_map="auto")
+    model = AutoModelForImageTextToText.from_pretrained(
+        model_name, device_map="auto", dtype=torch.bfloat16
+    )
     processor = AutoProcessor.from_pretrained(model_name)
     # Set left padding for decoder-only models
     processor.tokenizer.padding_side = "left"
